+++ react-workspace/README.md
@@ -4,7 +4,7 @@ This project was generated for React [v17.0.2](https://github.com/facebook/react
 
 This React Project is a utility wrapper to help build the `@trimble-oss/modus-react-components` library.
 
-All components in this repository are auto generated by the StencilJS library.
+All components in this repository are auto generated by the [Stencil](https://stenciljs.com/docs/introduction) library.
 
 To build a local copy of these components you can run:
