+++ react-workspace/package.json
@@ -42,7 +42,6 @@
     "@types/react-dom": "^17.0.11",
     "jest": "^28.1.1",
     "jest-dom": "^3.0.2",
-    "np": "^7.6.2",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "typescript": "^4.7.4"
