--- react-workspace/projects/react-tester/src/App.css
@@ -1,25 +1,0 @@
-.grid{
-  display: grid;
-  grid-template-columns: auto;
-  grid-gap: 10px;
-  padding: 10px;
-}
-
-.grid-row{
-  display: grid;
-  grid-auto-flow:column;
-  width:fit-content;
-  grid-column-gap: 10px;
-}
-
-.w-50{
-  width:50%;
-}
-
-
-h3{
-  color:white;
-  width:100%;
-  background: #0e416c;
-  padding: 10px;
-}
