+++ react-workspace/projects/test-react-v17/public/manifest.json
