+++ react-workspace/projects/test-react-v17/src/examples/ModusMessageExamples.tsx
@@ -1,0 +1,11 @@
+import { ModusMessage } from '@trimble-oss/modus-react-components';
+
+export default function ModusMessageExamples() {
+  return (
+    <>
+      <h3>Message</h3>
+      <ModusMessage type="info">Info message.</ModusMessage>
+      <ModusMessage type="question">Question message?</ModusMessage>
+    </>
+  );
+}
