+++ react-workspace/projects/test-react-v17/src/index.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './App';
+import '@trimble-oss/modus-web-components/dist/modus-web-components/modus-web-components.css';
 
 ReactDOM.render(
   <React.StrictMode>
