+++ react-workspace/tsconfig.json
@@ -10,7 +10,6 @@
     "noUnusedLocals": true,
     "noUnusedParameters": true,
     "removeComments": false,
-
     "outDir": "./dist",
     "lib": ["dom", "es2015"],
     "module": "es2015",
